+++ src/components/Banner/styled/StyledBanner.js
