+++ src/components/Footer/styled/StyledFooter.js
