+++ src/components/Form/Form.js
@@ -1,0 +1,8 @@
+﻿import React from 'react';
+import { Container } from './styled/StyledForm';
+
+const Form = ({ children, ...restProps }) => {
+  return <Container {...restProps}>{children}</Container>;
+};
+
+export default Form;
