+++ src/components/_old/Banner/Banner.js
@@ -1,0 +1,54 @@
+﻿/* eslint-disable camelcase */
+import React, { useState, useEffect } from 'react';
+import axios from 'axios/axios';
+import requests from 'axios/requests';
+import 'components/_old/Banner/Banner.scss';
+
+const Banner = () => {
+  const [movie, setMovie] = useState([]);
+
+  useEffect(() => {
+    async function fetchData() {
+      const request = await axios.get(requests.fetchNetflixOriginals);
+      setMovie(
+        request.data.results[
+          Math.floor(Math.random() * request.data.results.length - 1)
+        ],
+      );
+    }
+    fetchData();
+  }, []);
+
+  const truncate = (str, n) => {
+    return str?.length > n ? `${str.substr(0, n - 1)}...` : str;
+  };
+
+  return (
+    <header
+      className="banner"
+      style={{
+        backgroundImage: `url("https://image.tmdb.org/t/p/original/${movie?.backdrop_path}")`,
+      }}
+    >
+      <div className="banner__contents">
+        <h1 className="banner__title">
+          {movie?.name || movie?.title || movie?.original_name}
+        </h1>
+        <div className="banner__buttons">
+          <button type="button" className="banner__button">
+            Play
+          </button>
+          <button type="button" className="banner__button">
+            My List
+          </button>
+        </div>
+        <h2 className="banner__description">
+          {truncate(movie?.overview, 150)}
+        </h2>
+      </div>
+      <div className="banner--fadebottom" />
+    </header>
+  );
+};
+
+export default Banner;
