+++ src/components/_old/Banner/Banner.scss
