+++ src/components/index.js
@@ -1,0 +1,6 @@
+﻿export { default as Accordion } from 'components/Accordion/Accordion';
+export { default as Banner } from 'components/Banner/Banner';
+export { default as CTAForm } from 'components/CTAForm/CTAForm';
+export { default as Footer } from 'components/Footer/Footer';
+export { default as Header } from 'components/Header/Header';
+export { default as Jumbotron } from 'components/Jumbotron/Jumbotron';
