--- src/components/molecules/index.js
@@ -1,6 +1,0 @@
-﻿export { default as Footer } from 'components/molecules/Footer/Footer';
-export { default as Accordion } from 'components/molecules/Accordion/Accordion';
-export { default as Jumbotron } from 'components/molecules/Jumbotron/Jumbotron';
-export { default as CTAForm } from 'components/molecules/CTAForm/CTAForm';
-export { default as Header } from 'components/molecules/Header/Header';
-export { default as Banner } from 'components/molecules/Banner/Banner';
